Renamed the SingleRequestBidder to BidderFactory. Updated it to handle multi-request architecture adapters.
